Cover art
Using this protrait by an unknown artist, which has been dedicated to the public domain by the Met Museum: https://metmuseum.org/art/collection/search/437772

Selected on the basis that it’s a charismatic nose and probably the most famous Gogol short is The Nose (which is included in this collection). It also looks great in the crop.

I tried to find a PD source for https://commons.wikimedia.org/wiki/File:N.Gogol_by_F.Moller_(1840,_Tretyakov_gallery).jpg but without any luck. There’s a Gogol collection with a subtly different portrait but that one isn’t available as high quality. n.b. that’s marked as US PD on Wikimedia but without any justification, and I suspect the editor did it erroneously.